% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{bws}
\alias{bws}
\title{Function for running guided bootstrapping on wordscores algorithm using quanteda package}
\usage{
bws(
  text,
  doc_id = NULL,
  l_scale,
  r_scale,
  l_score = 1,
  r_score = 10,
  remove_no = FALSE,
  remove_punct = FALSE,
  tolower = FALSE,
  remove_sw = NULL,
  min_termfrq = NULL,
  max_termfrq = NULL
)
}
\arguments{
\item{text}{Vector of strings that is going to be scaled.}

\item{doc_id}{Vector of strings containing the labels for scaled documents.}

\item{l_scale}{Numeric vector for selecting documents that should be virtually placed on the left side of the scale. 'l_sacale' does not necessarily refer to ideological position, rather it denotes one of the virtual ends on the scale.}

\item{r_scale}{Numeric vector for selecting documents that should be virtually placed on the right side of the scale. 'r_sacale' does not necessarily refer to ideological position, rather it denotes one of the virtual ends on the scale.}

\item{l_score}{A number indicating virtual value for l_ side of the scale. It is one the Ys in quanteda's textmodel_wordscores function. Default is 1.}

\item{r_score}{A number indicating virtual value for r_ side of the scale. It is one the Ys in quanteda's textmodel_wordscores function. Default is 10.}

\item{remove_no}{Logical parameter for removing numbers. Default is FALSE.}

\item{remove_punct}{Logical parameter for removing punctuation. Default is FALSE.}

\item{tolower}{Logical parameter for transforming strings to lower case. Default is FALSE.}

\item{remove_sw}{specify language of stopwords by ISO 639-1 code. Default is NULL.}

\item{min_termfrq}{Minimum values of feature frequencies across all documents, below/above which features will be removed. Default is NULL.}

\item{max_termfrq}{Maximum values of feature frequencies across all documents, below/above which features will be removed. Default is NULL.}
}
\description{
This function allows training multiple wordscores models and extract stabilized indexes for scaled documents.
}
\examples{
bws()
}
\keyword{bootstrapping}
\keyword{scaling}
\keyword{wordscores}
