% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_viz.R
\name{bws_viz}
\alias{bws_viz}
\title{Function for visualizing scaled documents}
\usage{
bws_viz(bws_df, labels = NULL, se = "mean_se")
}
\arguments{
\item{bws_df}{A dataframe returned by bwordscores function.}

\item{labels}{Custom labels for scaled documents visualized in graph. Default is NULL.}

\item{se}{Type of standard errors used in the graph: "mean_se" or "se_ws". Default is mean_se.}
}
\description{
Simple function for visualizing wordscores of scaled documents.
}
\examples{
bws_viz()
}
\keyword{bootstrapping}
\keyword{scaling}
\keyword{visualizaiton}
\keyword{wordscores}
