% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws-package.R
\docType{package}
\name{bws-package}
\alias{bws-package}
\alias{bwspackage}
\title{The 'bws' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
