% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{bws}
\alias{bws}
\title{Bayesian Weighted Sums}
\usage{
bws(iter, y, X, Z = NULL, alpha = NULL, ...)
}
\arguments{
\item{iter}{Number of Hamiltonian Monte Carlo iterations}

\item{y}{Am n-vector of outcomes}

\item{X}{An n-by-p matrix of mixtures to be weightedly summed}

\item{Z}{Default NULL. A matrix of confounders whose linear effects are estimated}

\item{alpha}{A p-vector of hyperparameters for the Dirichlet prior on the weights.
Default to be a vector of 1's.}

\item{...}{Additional arguments for \code{rstan::sampling}}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
This function fits a Bayesian Weighted Sums as described in
Bayesian Weighted Sums: A Flexible Approach to Estimate Summed Mixture Effects.
Ghassan B. Hamra 1, Richard F. Maclehose, Lisa Croen, Elizabeth M. Kauffman
and Craig Newschaffer. 2021. International Journal of Environmental Research and Public Health
}
